requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic>=0.45.0
groq>=0.11.0
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
python-telegram-bot==21.10
praw==7.8.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
asyncpg==0.30.0
stripe==8.0.0
reportlab==4.2.0
matplotlib>=3.10.0
openai==1.50.0
twilio==9.0.0
redis[hiredis]==5.0.0
